feat: Add dynamic control for frame rendering
Functionality was included to allow dynamic control over when frames should be rendered. This change addresses the need for efficiency as it prevents unnecessary frame rendering when there are no updates. A boolean variable, do_render_this_frame_, was introduced to toggle the rendering each frame. This variable is checked before rendering to decide whether extensions for the frame begin, window render, and frame end are called.